requirements.txt
asgiref==3.4.0
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-url-filter==0.3.15
djangorestframework==3.12.4
drf-nested-routers==0.93.3
enum-compat==0.0.3
idna==2.10
lxml==4.6.3
pytz==2021.1
requests==2.25.1
route53==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
